refactor: extract detectors.rs from brain/insights.rs
Split insights.rs (1,223 lines) into 2 focused modules:
- detectors.rs (669 lines): 7 pure detection algorithms
- insights.rs (627 lines): types, state management, orchestration, display

Detection functions are pure (decisions in, insights out) with no I/O.
